% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bangladesh.R
\docType{data}
\name{bangladesh}
\alias{bangladesh}
\title{Country Boundary for Bangladesh
GADM is a spatial database of the location of the world's administrative areas (or adminstrative boundaries) for use in GIS and similar software. Administrative areas in this database are countries and lower level subdivisions such as provinces, departments, bibhag, bundeslander, daerah istimewa, fivondronana, krong, landsvæðun, opština, sous-préfectures, counties, and thana. GADM describes where these administrative areas are (the "spatial features"), and for each area it provides some attributes, such as the name and variant names. The current version is 2.8 (November 2015).}
\format{A spatialPolygonDataFrame}
\source{
\url{http://www.gadm.org/}
}
\usage{
bangladesh
}
\description{
Country Boundary for Bangladesh
GADM is a spatial database of the location of the world's administrative areas (or adminstrative boundaries) for use in GIS and similar software. Administrative areas in this database are countries and lower level subdivisions such as provinces, departments, bibhag, bundeslander, daerah istimewa, fivondronana, krong, landsvæðun, opština, sous-préfectures, counties, and thana. GADM describes where these administrative areas are (the "spatial features"), and for each area it provides some attributes, such as the name and variant names. The current version is 2.8 (November 2015).
}
\keyword{datasets}
