% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_dates.R
\docType{data}
\name{modis_dates}
\alias{modis_dates}
\title{Sample dates retrieved from the MODIS API with MODISTools package}
\format{character vector of dates in the MODIS format.}
\source{
\url{http://www.worldclim.org/bioclim}
}
\usage{
modis_dates
}
\description{
MODIS date codes: `.A2006001` - Julian Date of Acquisition (A-YYYYDDD)
To  Convert to a _proper_ date: 1)  Drop the "`A`" and 2)  Specify date format with julian day `[1,365]`
}
\keyword{datasets}
